@@ -12,12 +12,13 @@
     "license": "MIT",
     "homepage": "https://github.com/fetzi/server-timing",
     "require": {
-        "php": "^7.2",
+        "php": "^8.0",
         "psr/http-server-middleware": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^8.5",
-        "squizlabs/php_codesniffer": "^3.5"
+        "friendsofphp/php-cs-fixer": "^3.4",
+        "phpunit/phpunit": "^9.0",
+        "rector/rector": "^0.12.8"
     },
     "autoload": {
         "psr-4": {
@@ -31,7 +32,8 @@
     },
     "scripts": {
         "test": "vendor/bin/phpunit",
-        "lint": "vendor/bin/phpcs --standard=PSR2,PSR12 src/"
+        "lint": "php-cs-fixer fix -v --dry-run",
+        "fix": "php-cs-fixer fix -v"
     },
     "config": {
         "sort-packages": true
